Use Octohub pagination to get large numbers of releases.
Repositories with large numbers of releases would migrate only partially
because Octohub's Connection object does not follow pagination links.

Use Octohub's Pager object to fetch all releases. Flatten the nested
list of releases and use it for the migration.

This code is correct, but an outstanding bug in Octohub prevents it
from working correctly. A PR is submitted upstream.